Refactored login, signup and reset password pages
