+++ layout/ForgotPassword/components/index.ts
@@ -1,0 +1,3 @@
+import ForgotPasswordForm from "./ForgotPasswordForm";
+
+export { ForgotPasswordForm };
