+++ layout/ResetPassword/components/index.ts
@@ -1,0 +1,3 @@
+import ResetPasswordForm from "./ResetPasswordForm";
+
+export { ResetPasswordForm };
